backend-fastapi/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
PyJWT==2.9.0
pydantic==2.9.2
python-multipart==0.0.9
